Revert "[ Fix ] prefetch 적용 (#340)" (#341)
This reverts commit c4b9fa5c69c810f35fc51484ce5a9e66f7763ac6.
