+++ src/app/group/[groupId]/page.tsx
@@ -1,15 +1,13 @@
 import { getGroupInfo, getGroupMemberList } from "@/app/api/groups";
-import { getAllRanking, getTopRanking } from "@/app/api/groups/ranking";
+import { getTopRanking } from "@/app/api/groups/ranking";
 import { getDeadlineReachedProblems } from "@/app/api/problems";
 import { listSectionStyle, titleStyle } from "@/app/group/[groupId]/page.css";
 import Sidebar from "@/common/component/Sidebar";
 import ProblemList from "@/shared/component/ProblemList";
-import { prefetchQuery } from "@/shared/util/prefetch";
 import { sidebarWrapper } from "@/styles/shared.css";
 import GroupSidebar from "@/view/group/dashboard/GroupSidebar";
 import NoticeBanner from "@/view/group/dashboard/NoticeBanner";
 import Ranking from "@/view/group/dashboard/Ranking";
-import { HydrationBoundary } from "@tanstack/react-query";
 
 const GroupDashboardPage = async ({
   params: { groupId },
@@ -27,12 +25,6 @@ const GroupDashboardPage = async ({
       deadlineReachedData,
     ]);
 
-  const firstPage = 1;
-  const dehydratedState = await prefetchQuery({
-    queryKey: ["ranking", +groupId, firstPage],
-    queryFn: () => getAllRanking(+groupId, 0),
-  });
-
   return (
     <main className={sidebarWrapper}>
       <Sidebar>
@@ -40,9 +32,7 @@ const GroupDashboardPage = async ({
       </Sidebar>
       <div className={listSectionStyle}>
         <NoticeBanner />
-        <HydrationBoundary state={dehydratedState}>
-          <Ranking rankingData={rankingInfo} />
-        </HydrationBoundary>
+        <Ranking rankingData={rankingInfo} />
         <h2 className={titleStyle}>풀어야 할 문제</h2>
         <section>
           <ProblemList.Header />
