+++ src/shared/component/QueryProvider/index.tsx
@@ -3,24 +3,12 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { useState } from "react";
 
-const createQueryClient = () => {
-  return new QueryClient({
-    defaultOptions: {
-      queries: {
-        staleTime: 60 * 1000,
-      },
-    },
-  });
-};
-
-const QueryProvider = ({
+export default function QueryProvider({
   children,
 }: {
   children: React.ReactNode;
-}) => {
-  const [client] = useState(createQueryClient());
+}) {
+  const [client] = useState(new QueryClient());
 
   return <QueryClientProvider client={client}>{children}</QueryClientProvider>;
-};
-
-export default QueryProvider;
+}
